predict/requirements.txt
# Data processing
numpy>=1.26.0
pandas>=2.0.0
polars>=0.20.0  # Fast DataFrame library for large data

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Metrics and Evaluation
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Progress bars
tqdm>=4.66.0

# Logging
loguru>=0.7.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0